+++ lib/server/services/qualityControlFlag/QcFlagTypeService.js
@@ -46,7 +46,7 @@ class QcFlagTypesService {
      * @return {Promise<QcFlagType>} a Quality Control Flag Type
      */
     async getOneOrFail(id) {
-        const qcFlag = await this.getById(IDBDatabase);
+        const qcFlag = await this.getById(id);
         if (!qcFlag) {
             const criteriaExpression = `id (${id})`;
             throw new NotFoundError(`Quality Control Flag Type with this ${criteriaExpression} could not be found`);
